Switch to new TorchRec planner
Summary:
This diff replaces the legacy TorchRec planner with the new planner.

The new planner uses a new design and has some improvements, e.g. it fixes the shard merging issue in col-wise sharding. It has API input side changes while the output side remains the same.

This diff switches to the new planner API and the corresponding unit tests are changed as well.

Reviewed By: chrisxcai

Differential Revision: D33490573

fbshipit-source-id: 6ce7aefcb79878a0ce1f27c53b5b815f1d679e30